refactor: move server specific stuff out of game.ts
Game should now be free from nodejs imports other than config and logger...
will make rebasing offline mode easier and prevents unit tests from attempting network requests
